+++ Pong_Server/Racket.cpp
@@ -8,11 +8,11 @@ void Racket::Update(double DeltaTime)
 {
 	if (State == InputState::Up_Arrow)
 	{
-		transform->position.y -= (velocity * DeltaTime);
+		transform->position.y -= (velocity * (float) DeltaTime);
 	}
 	else if (State == InputState::Down_Arrow)
 	{
-		transform->position.y += (velocity * DeltaTime);
+		transform->position.y += (velocity * (float) DeltaTime);
 	}
 
 	ClampPosition();
